+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 ### Added
 
 * database: impl PartialEq and Eq for `Events` ([Yuki Kishimoto])
+* sdk: automatically resend event after NIP-42 authentication ([Yuki Kishimoto])
 
 ### Fixed